# Stabilize step_by by adding it to Iterator (issue #27741)