--- a/src/doc/unstable-book/src/library-features/iterator-step-by.md
+++ b/src/doc/unstable-book/src/library-features/iterator-step-by.md
@@ -1,0 +1,7 @@
+# `iterator_step_by`
+
+The tracking issue for this feature is: [#27741]
+
+[#27741]: https://github.com/rust-lang/rust/issues/27741
+
+------------------------