Inspired by itertools' `take()` method. See issue #27741