@ivandardi 2017-04-21
This is my first commit to Rust! :D ~~I probably screwed something up~~

I may need help with the docstrings. I feel like they can be improved. Also, I couldn't get it to compile without adding the `#[stable]` tags. Which are the proper tags to put there?