@ollie27 2017-04-21
How about if the number passed to `every_nth` is the number of elements that it should skip? It would mean `every_nth(0)` would return every element, `every_nth(1)` would return every other element and so on. It would match `nth` and remove the awkward `panic`.