@nagisa 2017-04-21
> How would every_nth(1) work?

Probably the first. You can get the behaviour of the 2nd code sample by skipping before hand, whereas you cannot get the first behaviour from the first.