@ivandardi 2017-04-22
Hm... It would be good to keep consistency with other languages too. If we keep the Python behavior, then the current implementation is correct and all it's needed is the `#[should_panic]` test.

After that we need to discuss whether or not to keep the name `every_nth` for clarity, how to fix the `#[stable]` tags and possibly the removal of the `Step` trait altogether.