@nagisa 2017-04-24
> If we keep the Python behavior, then the current implementation is correct

Too bad Python has a terrible behaviour in a lot of cases. I feel that at this point it is more in the garden of `T-libs` to decide what’s the optimal behaviour here.

Here’s what the documentation for the behaviour I would prefer:

``````
Returns an iterator adaptor which yields the next element and then skips the specified number of elements before yielding another element. `every_nth(0)` is an identity adaptor.

# Examples

```
let mut it = (0..).every_nth(0).take(2);
assert_eq!(it.next(), Some(0));
assert_eq!(it.next(), Some(1));
assert_eq!(it.next(), None);
```

```
let mut it = (0..).every_nth(2).take(3);
assert_eq!(it.next(), Some(0));
assert_eq!(it.next(), Some(3));
assert_eq!(it.next(), Some(6));
assert_eq!(it.next(), None);
```
``````

Implementation wise it also needs to be decided whether it is fine for iterator to be implemented as

```
fn next(&mut self) -> Option<Self::Item> { let n = self.next(); self.take(self.n).count(); n }
// as opposed to
fn next(&mut self) -> Option<Self::Item> { if self.first_take { self.next() } else { self.skip(self.n).next() } }
```

as these two have distinct behaviour in terms of the state the underlying iterator is after a `next`. Implementation approach will likely dictate the documentation wording as well.