@BurntSushi 2017-05-09
cc @rust-lang/libs Are we sufficiently past the `step_by` stuff that we'd be willing to take `every_nth`?