@BurntSushi 2017-05-09
@ivandardi The libs team discussed this today, and it seems like we're all on board with this direction. It looks like there are still some changes you wanted to make before merging though? For example, is it intentional that `every_nth(0)` panics?