@nagisa 2017-05-09
@BurntSushi I feel like there has been some mis-comunication? It was on the libs teams to decide on one of the many proposed behaviours, and I do not see any decision wrt that.