@BurntSushi 2017-05-09
@nagisa I see. I'll take a closer look and try to summarize.