@ivandardi 2017-05-09
Yes, it's expected that `every_nth(0)` panics, otherwise it would enter an infinite loop of yielding the first element forever. It's consistent with the behavior seen in Python too.

I'll submit a commit making the changes to the feature name and making it unstable hopefully later today.