@nagisa 2017-05-10
> Yes, it's expected that every_nth(0) panics, otherwise it would enter an infinite loop of yielding the first element forever. It's consistent with the behavior seen in Python too.

You did agree before that having `every_nth(0)` be identity iterator transformer could make sense. You don’t think so anymore?