@ivandardi 2017-05-10
@nagisa Nope. Makes more sense for `every_nth(1)` to be the identity, because it means that it will give every 1st element, aka the next element. It also goes in hand with what other languages have, and it's important to have familiar behavior. Yeah, I know I went back and forth in this, but this hasn't been easy to decide! XD But I think everyone is pretty settled on it now.

Also, I was testing with `step_by(0)` always yielding the first element of the iterable and I actually accidentally crashed my browser a lot of times in the playground while testing it out. If you have a `for i in (0..10).every_nth(0)`, it will loop forever unless you have a break inside the loop. Combine that with the fact that people might have something like

```rs
usize step = do_calculation()
for i in (0..len).step_by(step)
```
where `step` might accidentally be `0` and it's a really bad thing. I now see why Python throws an exception when the step parameter in their ranges is 0 and that's why I'm also making it panic on `step_by(0)`.