@ollie27 2017-05-10
> Makes more sense for every_nth(1) to be the identity, because it means that it will give every 1st element, aka the next element.

That doesn't match [`nth`](https://doc.rust-lang.org/nightly/std/iter/trait.Iterator.html#method.nth):

> Like most indexing operations, the count starts from zero, so nth(0) returns the first value, nth(1) the second, and so on.