@ivandardi 2017-05-10
That's why I think it's best to rename this to `step_by`. That way it doesn't make any correlation with `nth`.