@BurntSushi 2017-05-11
> Implementation wise it also needs to be decided whether it is fine for iterator to be implemented as

@nagisa Could you say more about how a caller could witness a difference in behavior between your two proposed implementation?