@quininer 2017-06-07
After I switch from `step_by` to `iterator_step_by`, the bench shows a serious performance degradation. https://github.com/quininer/newhope/commit/038aab77bbeca8e9273c375e59c471904c43ea78

before:
```
running 3 tests
test bench_newhope_keygen  ... bench:     117,488 ns/iter (+/- 21,484)
test bench_newhope_shareda ... bench:      49,293 ns/iter (+/- 17,389)
test bench_newhope_sharedb ... bench:     220,014 ns/iter (+/- 168,978)
```

after:
```
running 3 tests
test bench_newhope_keygen  ... bench:   1,189,949 ns/iter (+/- 353,381)
test bench_newhope_shareda ... bench:     542,612 ns/iter (+/- 80,311)
test bench_newhope_sharedb ... bench:   1,609,396 ns/iter (+/- 420,654)
```